refactor: move wait
